rename community.py to avoid name clash
