+++ networkx/algorithms/community/__init__.py
@@ -1,6 +1,6 @@
 from networkx.algorithms.community.asyn_lpa import *
 from networkx.algorithms.community.centrality import *
-from networkx.algorithms.community.generators import *
+from networkx.algorithms.community.community_generators import *
 from networkx.algorithms.community.kclique import *
 from networkx.algorithms.community.kernighan_lin import *
 from networkx.algorithms.community.quality import *
