+++ networkx/algorithms/community/community_generators.py
